[commit-patch] Add subversion support.
Hi David,

our developers were interested in using commit-patch to help with their
code commits to our subversion repository. I've created a patch that
adds support for subversion. We've been using this internally and
thought that you might be interested in the code addition. I've attached
the patch to this e-mail (gzipped to avoid issues with mail scanners).

Cheers,
Peter