[commit-patch] Use --relative with "git diff" so that commit-partial doesn't silently fail when you in a subdirectory.
"git apply" seems to filter the diff so that only files within the current
directory (and its descendants) get applied. It (crazily) does this silently
and there doesn't seem to be an option to override this.

The other option would be to cd to the project root just before calling "git
apply" but I think it's more intuitive to diff from your current
directory--just cd to the root yourself if you want to see the whole
project's changes.