Emacs interface no longer displays backend output as it comes in
The emacs interface creates and populates a *commit-patch* buffer. This is meant
to stay in the background unless something goes wrong. Due to a bug in Emacs
(http://debbugs.gnu.org/cgi/bugreport.cgi?bug=17815) this buffer was getting
displayed if using commit-patch with TRAMP, even when no error has occured.

This patch asks emacs to not update this buffer with the data as it comes in.
This is fine here, and fixes the issue.